// enums/store.ts
export enum StoreItems {
    BACKPACK = "Sauce Labs Backpack",
    BIKE_LIGHT = "Sauce Labs Bike Light",
    SHIRT = "Sauce Labs Bolt T-Shirt",
    JACKET = "Sauce Labs Fleece Jacket",
    ONESIE = "Sauce Labs Onesie",
    T_SHIRT = "Test.allTheThings() T-Shirt (Red)"
}

export enum StoreSortOptions {
    NAME_ASC = "az",
    NAME_DESC = "za",
    PRICE_ASC = "lohi",
    PRICE_DESC = "hilo"
}